break apart LQL parser helpers and removed use of `Logflare.DataCase` for test modules that do not need it
