+++ test/logflare/lql/backend_transformer_test.exs
@@ -1,5 +1,5 @@
 defmodule Logflare.Lql.BackendTransformerTest do
-  use Logflare.DataCase, async: true
+  use ExUnit.Case, async: true
 
   alias Logflare.Lql.BackendTransformer
   alias Logflare.Lql.BackendTransformer.BigQuery
